Refactor trigger-specific code in graph_path.rs into graph_path/operation.rs and graph_path/transition.rs
