+++ apollo-federation/src/query_graph/condition_resolver.rs
@@ -13,7 +13,7 @@ use crate::error::FederationError;
 use crate::operation::SelectionSet;
 use crate::query_graph::graph_path::ExcludedConditions;
 use crate::query_graph::graph_path::ExcludedDestinations;
-use crate::query_graph::graph_path::OpGraphPathContext;
+use crate::query_graph::graph_path::operation::OpGraphPathContext;
 use crate::query_graph::path_tree::OpPathTree;
 use crate::query_plan::QueryPlanCost;
 
@@ -184,7 +184,7 @@ impl ConditionResolverCache {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::query_graph::graph_path::OpGraphPathContext;
+    use crate::query_graph::graph_path::operation::OpGraphPathContext;
     //use crate::link::graphql_definition::{OperationConditional, OperationConditionalKind, BooleanOrVariable};
 
     #[test]
