+++ apollo-federation/src/query_plan/fetch_dependency_graph.rs
@@ -48,12 +48,12 @@ use crate::operation::VariableCollector;
 use crate::query_graph::QueryGraph;
 use crate::query_graph::QueryGraphEdgeTransition;
 use crate::query_graph::QueryGraphNodeType;
-use crate::query_graph::graph_path::OpGraphPathContext;
-use crate::query_graph::graph_path::OpGraphPathTrigger;
-use crate::query_graph::graph_path::OpPath;
-use crate::query_graph::graph_path::OpPathElement;
-use crate::query_graph::graph_path::concat_op_paths;
-use crate::query_graph::graph_path::concat_paths_in_parents;
+use crate::query_graph::graph_path::operation::OpGraphPathContext;
+use crate::query_graph::graph_path::operation::OpGraphPathTrigger;
+use crate::query_graph::graph_path::operation::OpPath;
+use crate::query_graph::graph_path::operation::OpPathElement;
+use crate::query_graph::graph_path::operation::concat_op_paths;
+use crate::query_graph::graph_path::operation::concat_paths_in_parents;
 use crate::query_graph::path_tree::OpPathTree;
 use crate::query_graph::path_tree::PathTreeChild;
 use crate::query_plan::FetchDataPathElement;
