+++ apollo-federation/src/query_plan/query_planning_traversal.rs
@@ -21,17 +21,17 @@ use crate::query_graph::condition_resolver::ConditionResolution;
 use crate::query_graph::condition_resolver::ConditionResolutionCacheResult;
 use crate::query_graph::condition_resolver::ConditionResolver;
 use crate::query_graph::condition_resolver::ConditionResolverCache;
-use crate::query_graph::graph_path::ClosedBranch;
-use crate::query_graph::graph_path::ClosedPath;
 use crate::query_graph::graph_path::ExcludedConditions;
 use crate::query_graph::graph_path::ExcludedDestinations;
-use crate::query_graph::graph_path::OpGraphPath;
-use crate::query_graph::graph_path::OpGraphPathContext;
-use crate::query_graph::graph_path::OpPathElement;
-use crate::query_graph::graph_path::OpenBranch;
-use crate::query_graph::graph_path::SimultaneousPaths;
-use crate::query_graph::graph_path::SimultaneousPathsWithLazyIndirectPaths;
-use crate::query_graph::graph_path::create_initial_options;
+use crate::query_graph::graph_path::operation::ClosedBranch;
+use crate::query_graph::graph_path::operation::ClosedPath;
+use crate::query_graph::graph_path::operation::OpGraphPath;
+use crate::query_graph::graph_path::operation::OpGraphPathContext;
+use crate::query_graph::graph_path::operation::OpPathElement;
+use crate::query_graph::graph_path::operation::OpenBranch;
+use crate::query_graph::graph_path::operation::SimultaneousPaths;
+use crate::query_graph::graph_path::operation::SimultaneousPathsWithLazyIndirectPaths;
+use crate::query_graph::graph_path::operation::create_initial_options;
 use crate::query_graph::path_tree::OpPathTree;
 use crate::query_plan::QueryPlanCost;
 use crate::query_plan::fetch_dependency_graph::FetchDependencyGraph;
